raspberry_pi_receiver/requirements.txt
psycopg2-binary==2.9.9
RPi.GPIO==0.7.1
spidev==3.6
pyLoRa==0.4.0
